+++ app/app.js
@@ -1,15 +1,12 @@
 'use strict';
 
-angular.module('app').config(function($locationProvider, $stateProvider, $urlRouterProvider) {
+angular.module('app').config(function($locationProvider, $routeProvider) {
   $locationProvider.html5Mode(true);
 
-  $stateProvider.state('index', {
-    url: '/',
-    templateUrl: 'templates/index.html'
-  });
-
-  $urlRouterProvider.otherwise('/');
-})
-.run(function($rootScope) {
-  $rootScope.title = 'Aang Brunch';
+  $routeProvider
+  .when('/', {
+    templateUrl: 'templates/index.html',
+    controller: 'Index'
+  })
+  .otherwise('/');
 });