+++ app/controllers/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+angular.module('app.controllers').controller('Index', function($scope) {
+  $scope.title = 'Aang Brunch';
+});